Move TensorData.*pp to H5
