+++ src/DataStructures/Python/Bindings.cpp
@@ -8,11 +8,9 @@ namespace py = pybind11;
 namespace py_bindings {
 void bind_datavector(py::module& m);  // NOLINT
 void bind_matrix(py::module& m);      // NOLINT
-void bind_tensordata(py::module& m);  // NOLINT
 }  // namespace py_bindings
 
 PYBIND11_MODULE(_PyDataStructures, m) {  // NOLINT
   py_bindings::bind_datavector(m);
   py_bindings::bind_matrix(m);
-  py_bindings::bind_tensordata(m);
 }
