+++ src/DataStructures/Tensor/CMakeLists.txt
@@ -1,12 +1,6 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-spectre_target_sources(
-  ${LIBRARY}
-  PRIVATE
-  TensorData.cpp
-  )
-
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
@@ -19,7 +13,6 @@ spectre_target_headers(
   Structure.hpp
   Symmetry.hpp
   Tensor.hpp
-  TensorData.hpp
   TypeAliases.hpp
   )
