+++ src/IO/H5/CMakeLists.txt
@@ -19,6 +19,7 @@ spectre_target_sources(
   SourceArchive.cpp
   SpectralIo.cpp
   StellarCollapseEos.cpp
+  TensorData.cpp
   Version.cpp
   VolumeData.cpp
   )
@@ -40,6 +41,7 @@ spectre_target_headers(
   SourceArchive.hpp
   SpectralIo.hpp
   StellarCollapseEos.hpp
+  TensorData.hpp
   Type.hpp
   Version.hpp
   VolumeData.hpp
