+++ tests/Unit/IO/H5/CMakeLists.txt
@@ -11,6 +11,7 @@ set(LIBRARY_SOURCES
   Test_H5File.cpp
   Test_OpenGroup.cpp
   Test_StellarCollapseEos.cpp
+  Test_TensorData.cpp
   Test_Version.cpp
   Test_VolumeData.cpp
   )
@@ -53,5 +54,12 @@ spectre_add_python_bindings_test(
   "Unit.IO.H5.Python.ExtractDatFromH5"
   "Python/Test_ExtractDatFromH5.py"
   "unit;DataStructures;IO;H5;python"
+  PyH5
+  )
+
+spectre_add_python_bindings_test(
+  "Unit.IO.H5.Python.TensorData"
+  "Test_TensorData.py"
+  "unit;IO;H5;Python"
   PyH5
   )
